# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/banners": {
            "put": {
                "description": "Update a banner with new data and optionally a new file",
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "banners"
                ],
                "summary": "Update a banner",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Banner data",
                        "name": "banner",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Expiration date",
                        "name": "expires_at",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "File",
                        "name": "file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Failed to read file",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to update banner",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/projects/{project_id}/banners": {
            "post": {
                "description": "Create a new banner and save the image",
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "banners"
                ],
                "summary": "Create a new banner",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Story ID",
                        "name": "story_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "File",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Langs",
                        "name": "langs",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Failed to read file",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to unmarshal banner",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/projects/{project_id}/stories": {
            "get": {
                "description": "Get stories by project ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stories"
                ],
                "summary": "Get stories",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Story ID",
                        "name": "story_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Banner ID",
                        "name": "banner_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Creator",
                        "name": "creator",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Language",
                        "name": "lang",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "All",
                        "name": "all",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stories",
                        "schema": {
                            "$ref": "#/definitions/transport.GetStoriesResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get stories",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/views": {
            "post": {
                "description": "Create a new view for a banner",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "views"
                ],
                "summary": "Create a new view",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Banner ID",
                        "name": "banner_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to insert view",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "transport.GetStoriesResponse": {
            "type": "object",
            "properties": {
                "stories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Story"
                    }
                }
            }
        },
        "types.Banner": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "langs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.BannerLang"
                    }
                },
                "media_url": {
                    "type": "string"
                },
                "views": {
                    "type": "integer"
                }
            }
        },
        "types.BannerLang": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "lang": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "types.Story": {
            "type": "object",
            "properties": {
                "banners": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Banner"
                    }
                },
                "created_at": {
                    "type": "integer"
                },
                "creator": {
                    "type": "integer"
                },
                "expires_at": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                }
            }
        }
    }
}